requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
decouple==0.0.7
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
kombu==5.5.4
Markdown==3.9
packaging==25.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
